refactor: 💡 moved DotLottieSource to utils and updated readme
